+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -10,6 +10,10 @@
   [Grpc.AspNetCore](https://www.nuget.org/packages/Grpc.AspNetCore/).
   This option is enabled by default.
   ([#1423](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1423))
+* Renamed TextMapPropagator to TraceContextPropagator, CompositePropapagor
+  to CompositeTextMapPropagator. IPropagator is renamed to TextMapPropagator
+  and changed from interface to abstract class.
+  ([#1427](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1427))
 
 ## 0.7.0-beta.1